refactor central app logic and let view show data
